refactor(signals): share report_inbox_url helper and wire into autostart PR prompt
Move the canonical inbox deep-link builder into a lightweight `report_links` module (avoids a circular import via the Temporal-importing `utils` module) and route the existing URL-builder sites through it: the autostart implementation-task prompt (so the autonomous PR footer link and the serializer's report_url can't drift), the Slack inbox notifications, scout Slack delivery, and the scout report tool.

The serializer's report_url field now sources from the same helper.

Generated-By: PostHog Code
Task-Id: d4aa49fb-064c-4949-8c88-5f3b483a1c6f
